move admin commands to a separate module
