+++ app/controllers/concerns/admin_commands.rb
@@ -1,0 +1,23 @@
+module AdminCommands
+  extend ActiveSupport::Concern
+
+  included do
+    before_action :authorize_admin, only: %i[random_script! script_templates! admin!]
+  end
+
+  def random_script!(*)
+    ScriptGenerator::GenerateRandomScriptJob.perform_async(chat["id"])
+
+    respond_with :message, text: "Started script generation."
+  end
+
+  def script_templates!(*)
+    ScriptGenerator::SendScriptTemplatesJob.perform_async(chat["id"])
+
+    respond_with :message, text: "Fetching script templates."
+  end
+
+  def admin!(*)
+    respond_with :message, text: I18n.t("telegram_webhooks.commands.admin")
+  end
+end
