+++ app/controllers/telegram_webhooks_controller.rb
@@ -1,11 +1,10 @@
 class TelegramWebhooksController < Telegram::Bot::UpdatesController
   include Telegram::Bot::UpdatesController::MessageContext
+  include AdminCommands
   include TelegramLocale
   include SessionAccessor
   include ErrorHandler
   include TgChatAuthorization
-
-  before_action :authorize_admin, only: %i[random_script! script_templates! admin!]
 
   def start!(token_code = nil)
     handled_token = TokenHandler::HandleToken.call(
@@ -93,22 +92,6 @@ class TelegramWebhooksController < Telegram::Bot::UpdatesController
       tg_message_id:,
       callback_query_id:
     )
-  end
-
-  def random_script!(*)
-    ScriptGenerator::GenerateRandomScriptJob.perform_async(chat["id"])
-
-    respond_with :message, text: "Started script generation."
-  end
-
-  def script_templates!(*)
-    ScriptGenerator::SendScriptTemplatesJob.perform_async(chat["id"])
-
-    respond_with :message, text: "Fetching script templates."
-  end
-
-  def admin!(*)
-    respond_with :message, text: I18n.t("telegram_webhooks.commands.admin")
   end
 
   private
